requirements.txt
openai>=1.0.0
dspy-ai>=2.0.0
deepeval>=0.21.0
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pytest>=8.0.0
